recomendation sytem
In this project, I created a basic recommendation system by suggesting items that are most similar to a particular item, in this case, movies. This is not a true robust recommendation system, to describe it more accurately,it just tells what movies/items are most similar to a given movie choice.